Fix subdirectory bug
In the original 0.6.1 code there's a for-loop involving a linked list. It triggers after a new directory is found to see if any directories in the unknown list are related to it. When a match is found, the link ->next is changed to null and the loop ends prematurely. This while-loop preserves the address of the next object in case it has been clobbered.